[Model Registry] Rename unity_catalog_oss_* protos to unity_catalog_*
Drop the "oss_" prefix from the shared UC governance proto file names per review
feedback: unity_catalog_oss_{messages,service}.proto ->
unity_catalog_{messages,service}.proto (and the generated _pb2.py/.pyi).

Updates the internal service->messages proto import, dev/generate_protos.py
(file list + gencode-rewrite rule), the dev/pyproject.py and libs/tracing
exclude lists, and all import references in the consuming client/util/test
modules. Regenerated the gencode so the serialized descriptors carry the new
file names.
