+++ mlflow/protos/unity_catalog_service.proto
@@ -4,7 +4,7 @@ package mlflow.unitycatalog;
 
 import "databricks.proto";
 import "scalapb/scalapb.proto";
-import "unity_catalog_oss_messages.proto";
+import "unity_catalog_messages.proto";
 
 option java_generate_equals_and_hash = true;
 option java_package = "com.databricks.api.proto.managedcatalog";
