+++ tests/utils/test_unity_catalog_oss_utils.py
@@ -1,5 +1,5 @@
 from mlflow.entities.model_registry import RegisteredModel
-from mlflow.protos.unity_catalog_oss_messages_pb2 import (
+from mlflow.protos.unity_catalog_messages_pb2 import (
     RegisteredModelInfo,
 )
 from mlflow.utils._unity_catalog_oss_utils import get_registered_model_from_uc_oss_proto
